Remove default values from offset and key gen to live in cipher class
