+++ lib/key_gen.rb
@@ -3,12 +3,6 @@ require_relative './modules/create_defaults'
 class KeyGen
   include Generable
 
-  attr_reader :num
-
-  def initialize(num = generate_random_key)
-    @num = num
-  end
-
   def create_key(five_digit_key)
     { key_A: five_digit_key[0..1].to_i,
       key_B: five_digit_key[1..2].to_i,
