+++ lib/modules/create_defaults.rb
@@ -6,4 +6,12 @@ module Generable
     end
     rand_num
   end
+
+  def create_character_set
+    ("a".."z").to_a << " "
+  end
+
+  def create_date
+    Date.today.strftime('%d%m%y')
+  end
 end
